requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
idna==2.10
more-itertools==8.6.0
pycodestyle==2.6.0
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.5
requests==2.25.1
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
